requirements.txt
Flask
gunicorn
requests
spotipy
python-dotenv